Online movement and attack

+++ core/src/com/sticksouls/items/weapons/Weapon.java
@@ -17,6 +17,7 @@ import com.badlogic.gdx.physics.box2d.World;
 import com.badlogic.gdx.physics.box2d.joints.RevoluteJoint;
 import com.badlogic.gdx.physics.box2d.joints.RevoluteJointDef;
 import com.badlogic.gdx.utils.Array;
+import com.sticksouls.enums.Directions;
 import com.sticksouls.items.Item;
 import com.sticksouls.utils.Resources;
 
@@ -28,6 +29,7 @@ public abstract class Weapon extends Item {
 	protected TextureRegion[][] empoweredAttackFrames;
 	protected Animation<TextureRegion> basicAttackAnimation;
 	protected Animation<TextureRegion> empoweredAttackAnimation;
+	protected Directions direction;
 	
 	protected final Body CHARACTERBODY;
 	private World world;
@@ -37,7 +39,7 @@ public abstract class Weapon extends Item {
 	private int height;
 	protected boolean firstDraw = false;
 	
-	protected final RevoluteJoint rightJoint, leftJoint, topJoint, bottomJoint;
+	protected RevoluteJoint rightJoint, leftJoint, topJoint, bottomJoint;
 	protected RevoluteJointDef rightJointDef, leftJointDef, topJointDef, bottomJointDef;
 	private boolean isRightJoint, isLeftJoint, isTopJoint, isBottomJoint;
 	
@@ -63,8 +65,9 @@ public abstract class Weapon extends Item {
 		topJoint = createJoint(new Vector2(0, 8), new Vector2(0, -15), topJointDef);
 		bottomJoint = createJoint(new Vector2(0, -8), new Vector2(0, -15), bottomJointDef);
 		
-		destroyAllJoints();
-		createJoint(rightJointDef);
+		destroyAllMyJoints(null);
+		
+		rightJoint = createJoint(rightJointDef);
 		
 	}
 
@@ -79,22 +82,26 @@ public abstract class Weapon extends Item {
 		return (RevoluteJoint) world.createJoint(jointDef);
 	}
 	
-	protected void createJoint(RevoluteJointDef joint) {
-		world.createJoint(joint);
+	protected RevoluteJoint createJoint(RevoluteJointDef jointDef) {
+		return (RevoluteJoint) world.createJoint(jointDef);
 	}
 	
-	protected void destroyAllJoints() {
+	protected void destroyAllMyJoints(RevoluteJoint jointToDestroy) {
 		Array<Joint> joints = new Array<Joint>();
 		world.getJoints(joints);
 		
 		for(Joint joint : joints) {
-			world.destroyJoint(joint);	
+			if(joint.equals(topJoint) || joint.equals(rightJoint) || joint.equals(bottomJoint) || joint.equals(leftJoint) || joint.equals(jointToDestroy)) {
+				world.destroyJoint(joint);						
+			}			
 		}
+		
 	}
 	
 	public void draw() {
 		if(!firstDraw) {
 			firstDraw = true;
+			
 			weaponBody.setTransform(weaponBody.getPosition().x, weaponBody.getPosition().y, 0);
 		}
 		
@@ -129,6 +136,7 @@ public abstract class Weapon extends Item {
 	}
 
 	public abstract void attack(Vector2 characterCoordinates);
+	public abstract void attack(Directions direction);
 	public abstract boolean continueAttack();
 	protected abstract boolean finishAttack();
 	
@@ -144,5 +152,9 @@ public abstract class Weapon extends Item {
 	public int getEmpoweredDmg() {
 		return empoweredDmg;
 	}	
+	
+	public Directions getDirection() {
+		return direction;
+	}
 }
